requirements.txt
python-dotenv
flask
flask-cors
openai
PyPDF2
dirtyjson
gunicorn
requests